:recycle:refactor analysis text
