+++ src/main/java/com/uio/java_tools/controller/TokenizerController.java
@@ -1,13 +1,16 @@
 package com.uio.java_tools.controller;
 
 import com.uio.java_tools.common.BackMessage;
-import com.uio.java_tools.dto.EntityParameterDTO;
-import com.uio.java_tools.service.impl.TokenizerServiceImpl;
-import com.uio.java_tools.resq.StringDataReq;
+import com.uio.java_tools.dto.AnalysisDTO;
+import com.uio.java_tools.controller.req.StringDataReq;
+import com.uio.java_tools.enums.AnalysisTypeEnum;
+import com.uio.java_tools.service.TokenizerService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+
+import java.util.Map;
 
 
 /**
@@ -19,18 +22,24 @@ import org.springframework.web.bind.annotation.RestController;
 public class TokenizerController {
 
     @Autowired
-    private TokenizerServiceImpl tokenizerService;
+    private Map<String, TokenizerService> tokenizerServiceMap;
 
+    private static final String TOKENIZER_PREFIX = "tokenizer_";
     /**
      * 文本识别接口
      * @param dateMessage String data
-     * @return 识别好的List<String>
+     * @return analysisDTO
      */
-    @PostMapping(value = "/wordIdentify")
-    public BackMessage<EntityParameterDTO> tokenizer(@RequestBody StringDataReq dateMessage) {
-        // 1. 修改为解析成封装好的特定对象，键接口返回值
-        // 2. 新建返回值对象，封装字段名、类型和备注对象为成员，参照接口文档
-        EntityParameterDTO entityParameterDTO = tokenizerService.parseJavaEntityCode(dateMessage.getData());
-        return BackMessage.success(entityParameterDTO);
+    @PostMapping(value = "/analysisText")
+    public BackMessage<AnalysisDTO> analysisText(@RequestBody StringDataReq dateMessage) {
+
+        AnalysisTypeEnum analysisTypeEnum = AnalysisTypeEnum.valueOf(dateMessage.getAnalysisType());
+        TokenizerService tokenizerService = tokenizerServiceMap.get(TOKENIZER_PREFIX + analysisTypeEnum.getDesc());
+        if (null != tokenizerService) {
+            // 修改为解析成封装好的特定对象
+            AnalysisDTO analysisDTO = tokenizerService.analysisText(dateMessage.getData());
+            return BackMessage.success(analysisDTO);
+        }
+        return BackMessage.success(null);
     }
 }
