+++ src/main/java/com/uio/java_tools/dto/AnalysisJavaDTO.java
@@ -1,0 +1,14 @@
+package com.uio.java_tools.dto;
+
+/**
+ * @author han xun
+ * Date 2021/11/6 10:30
+ * Description:
+ */
+public class AnalysisJavaDTO extends AnalysisDTO {
+
+    /**
+     * 全路径名
+     */
+    private String namespace;
+}
