+++ src/main/java/com/uio/java_tools/dto/JavaParseParameterDTO.java
@@ -1,17 +1,17 @@
-package com.uio.java_tools.dto;
-
-import lombok.Data;
-
-/**
- * @author han xun
- * Date 2021/6/26 16:19
- * Description:
- */
-@Data
-public class JavaParseParameterDTO extends ParseParameterDTO{
-
-    /**
-     * 包名
-     */
-    private String packageName;
-}
+//package com.uio.java_tools.dto;
+//
+//import lombok.Data;
+//
+///**
+// * @author han xun
+// * Date 2021/6/26 16:19
+// * Description:
+// */
+//@Data
+//public class JavaParseParameterDTO extends ParseParameterDTO{
+//
+//    /**
+//     * 包名
+//     */
+//    private String packageName;
+//}
