+++ src/main/java/com/uio/java_tools/dto/ParseParameterDTO.java
@@ -1,27 +1,27 @@
-package com.uio.java_tools.dto;
-
-import lombok.Data;
-
-import java.util.List;
-
-/**
- * @author han xun
- * Date 2021/6/26 16:17
- * Description:
- */
-@Data
-public class ParseParameterDTO {
-
-    /**
-     * 字段列表
-     */
-    private List<String> fieldList;
-    /**
-     * 类名
-     */
-    private String className;
-    /**
-     * 表名
-     */
-    private String tableName;
-}
+//package com.uio.java_tools.dto;
+//
+//import lombok.Data;
+//
+//import java.util.List;
+//
+///**
+// * @author han xun
+// * Date 2021/6/26 16:17
+// * Description:
+// */
+//@Data
+//public class ParseParameterDTO {
+//
+//    /**
+//     * 字段列表
+//     */
+//    private List<String> fieldList;
+//    /**
+//     * 类名
+//     */
+//    private String className;
+//    /**
+//     * 表名
+//     */
+//    private String tableName;
+//}
