+++ src/main/java/com/uio/java_tools/service/GeneratorService.java
@@ -1,36 +1,33 @@
-package com.uio.java_tools.service;
-
-import com.uio.java_tools.dto.JavaParseParameterDTO;
-import com.uio.java_tools.dto.ParameterDTO;
-import com.uio.java_tools.dto.ParseParameterDTO;
-
-/**
- * @author han xun
- * Date 2021/6/26 16:23
- * Description: 生成文件服务
- */
-public interface GeneratorService {
-
-
-    /**
-     * 解析输入的字符串
-     *
-     * @param inputString
-     * @return
-     */
-    <T extends ParseParameterDTO> T parse(String inputString);
-
-    /**
-     * 生成DAO文件
-     */
-    default void generatorDAO() {
-
-    }
-
-    /**
-     * 生成XML文件
-     */
-    default void generatorXML() {
-
-    }
-}
+//package com.uio.java_tools.service;
+//
+//
+///**
+// * @author han xun
+// * Date 2021/6/26 16:23
+// * Description: 生成文件服务
+// */
+//public interface GeneratorService {
+//
+//
+//    /**
+//     * 解析输入的字符串
+//     *
+//     * @param inputString
+//     * @return
+//     */
+//    <T extends ParseParameterDTO> T parse(String inputString);
+//
+//    /**
+//     * 生成DAO文件
+//     */
+//    default void generatorDAO() {
+//
+//    }
+//
+//    /**
+//     * 生成XML文件
+//     */
+//    default void generatorXML() {
+//
+//    }
+//}
