+++ src/main/java/com/uio/java_tools/service/impl/MybatisConvertServiceImpl.java
@@ -3,6 +3,7 @@ package com.uio.java_tools.service.impl;
 import com.uio.java_tools.service.MybatisConvertService;
 import com.uio.java_tools.manager.impl.MybatisProduceManagerImpl;
 import com.uio.java_tools.dto.ParameterDTO;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -12,6 +13,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 @Service
+@Slf4j
 public class MybatisConvertServiceImpl implements MybatisConvertService {
 
     @Autowired
