+++ src/main/java/com/uio/java_tools/service/impl/SQLConvertServiceImpl.java
@@ -1,13 +1,13 @@
 package com.uio.java_tools.service.impl;
 
 
-import com.uio.java_tools.dto.EntityParameterDTO;
+import com.uio.java_tools.controller.req.CreateSqlReq;
+import com.uio.java_tools.dto.AnalysisDTO;
 import com.uio.java_tools.dto.Parameter;
-import com.uio.java_tools.common.BackEnum;
+import com.uio.java_tools.dto.ParameterDTO;
 import com.uio.java_tools.manager.impl.ParseStrManagerImpl;
 import com.uio.java_tools.manager.impl.VelocityTemplateForSQL;
 import com.uio.java_tools.service.SQLConvertService;
-import com.uio.java_tools.dto.ParameterDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -24,28 +24,28 @@ public class SQLConvertServiceImpl implements SQLConvertService {
     private VelocityTemplateForSQL velocityTemplateForSQL;
 
     /**
+     *
      * 创建数据库命令
-     * @param parameterDTO parameter,keyParameter,tbName
+     * @param createSqlReq
      * @return 创建数据库命令
      */
     @Override
-    public String createSqlService(EntityParameterDTO parameterDTO) {
-        if (parameterDTO.isUnderline()) {
-            for (Parameter parameter: parameterDTO.getParameters()) {
-                parameter.setField(parse.upperToLower(parameter.getField()));
-            }
+    public String createSqlService(CreateSqlReq createSqlReq) {
+
+        for (Parameter parameter: createSqlReq.getParameterList()) {
+            parameter.setField(parse.upperToLower(parameter.getField()));
         }
 
-        String SQLCommand = velocityTemplateForSQL.createSQLTemplate(parameterDTO.getParameters(), parameterDTO.getTableName(), parameterDTO.getPrimaryKey());
-        return SQLCommand;
+        return velocityTemplateForSQL.createSQLTemplate(createSqlReq.getParameterList(), createSqlReq.getTableName(), createSqlReq.getPrimaryKey());
     }
 
+
     /**
+     *
      * 更新数据库表命令
      * @param parameterDTO parameter,keyParameter,tbName
      * @return 更新数据库命令
      */
-
     @Override
     public String updateTableService(ParameterDTO parameterDTO) {
         // 需求改的参数
@@ -71,6 +71,7 @@ public class SQLConvertServiceImpl implements SQLConvertService {
     }
 
     /**
+     *
      * 插入数据库命令
      * @param parameterDTO parameter，tbName
      * @return 插入数据库命令
@@ -90,6 +91,7 @@ public class SQLConvertServiceImpl implements SQLConvertService {
     }
 
     /**
+     *
      * 生成删除信息命令
      * @param parameterDTO keyParameter,tbName
      * @return 删除信息命令
@@ -110,8 +112,8 @@ public class SQLConvertServiceImpl implements SQLConvertService {
         return deleteSQL;
     }
 
-
     /**
+     *
      * 查询信息命令生成
      * @param parameterDTO parameter，keyParameter，tbName
      * @return 查询数据库信息命令
