+++ src/test/java/com/uio/java_tools/service/GeneratorByJavaServiceTest.java
@@ -1,6 +1,6 @@
+/*
 package com.uio.java_tools.service;
 
-import com.uio.java_tools.dto.JavaParseParameterDTO;
 import com.uio.java_tools.utils.FileUtils;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,11 +9,13 @@ import org.springframework.boot.test.context.SpringBootTest;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
+*/
 /**
  * @author han xun
  * Date 2021/6/27 01:16
  * Description:
- */
+ *//*
+
 @SpringBootTest
 public class GeneratorByJavaServiceTest {
 
@@ -21,14 +23,18 @@ public class GeneratorByJavaServiceTest {
     @Qualifier("generatorByJavaServiceImpl")
     private GeneratorService generatorService;
 
-    /**
+    */
+/**
      * 解析出的数据应有的数量
-     */
+     *//*
+
     private static final int listSize = 9;
 
-    /**
+    */
+/**
      * 待测试文本存储路径
-     */
+     *//*
+
     private static String filePath = "target/classes/static/testJava.txt";
 
     @Test
@@ -42,3 +48,4 @@ public class GeneratorByJavaServiceTest {
 
 
 }
+*/
